pyproject.toml
[tool.poetry]
name = "s2ctl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]


[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.7.2"
click = "^7.1.2"
tabulate = "^0.8.7"
keyring = "^21.5.0"
config-path = "^1.0.0"
"keyrings.cryptfile" = "^1.3.6"
async-timeout = "^3.0.1"
PyYAML = "^5.3.1"
click-completion = "^0.5.2"
certifi = "^2020.12.5"

[tool.poetry.dev-dependencies]
tox = "^3.20.1"
flake8 = "^3.8.4"
wemake-python-styleguide = "^0.14.1"
pytest = "^6.1.2"
pytest-xdist = "^2.1.0"
pytest-aiohttp = "^0.3.0"
pytest-cov = "^2.10.1"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
s2ctl = 's2ctl.entrypoint:run_cli'

[pytest]
minversion = "6.0"
addopts = "-ra -q -n8"
testpaths = [
    "tests",
]

[tool.isort]
multi_line_output = 3
line_length = 120
include_trailing_comma = true
